requirements.txt
PyYAML
requests
flask